requirements.txt
krakenex
pandas
python-dotenv
SQLAlchemy
pymysql
ta
schedule
streamlit
streamlit-autorefresh
pushover-python
schedule
ta 
